+++ tests/index.test.ts
@@ -1,4 +1,4 @@
-import { expect, test, describe, it } from "vitest";
+import { expect, test, describe } from "vitest";
 
 import * as api from "../src";